backend/ocr-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy==1.24.3
pdf2image==1.16.3
python-dotenv==1.0.0
pydantic==2.5.0
google-cloud-vision==3.7.2
